+++ src/components/attendanceAdmin/totalScore/MemberList/style.ts
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+
+export const StListHeader = styled.header`
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 5rem;
+
+  & > h1 {
+    font-weight: 600;
+    font-size: 20px;
+    line-height: 25px;
+    letter-spacing: -0.02em;
+  }
+`;
